+++ packages/coverde_cli/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **BREAKING FEAT**: full support for analyzer package v8-v10 (#263).
 - **FEAT**: add `transform` command (#276).
+- **FEAT**: support transformation presets in `coverde.yaml` to be used in `transform` command (#278).
 
 ## 0.3.0+1